make sure to use cached meta data for parsing instead of previous (#33)
If a query is prepared the metadata for the given parameters is cached.
If before the query is executed with the same parameter types again,
another query is executed, the metadata in the connection is still
set to the metadata of that query.
It isn't updated since the metadata is assumed to be cached.
This fixes that by explicitly using the cached metadata in that case.