Basic support for transactions
With this change, it is possible to start, commit and rollback
transactions using the `simple_query` method in the `Client`. Querying
`BEGIN TRAN` will start a new transaction and store the current
transaction descriptor, adding it to the subsequent queries until
`COMMIT` or `ROLLBACK` happens.

Please be careful on always commiting or rolling back before subsequent
queries to always keep the transaction descriptor in a correct value.

This commit is a preliminary setup for allowing to use transactions.
Transaction management and automatic rollbacks should be implemented in
a separate pull request.